requirements.txt
opencv-python
scikit-learn
tensorflow
pandas
numpy
matplotlib.pyplot
streamlit
